requirements.txt
treamlit
pandas
numpy
scikit-learn
seaborn
matplotlib
torch